+++ backend/src/schemas/notification.schema.ts
@@ -3,8 +3,21 @@ import { HydratedDocument } from 'mongoose';
 import { ApiProperty } from '@nestjs/swagger';
 import { Expose } from 'class-transformer';
 import { randomUUID } from 'crypto';
+import { LanguageEnum } from './language.enum';
 
 export type NotificationDocument = HydratedDocument<NotificationSession>;
+
+/**
+ * Генерирует схему полей для каждого языка в enum,
+ * каждое поле — обязательная строка.
+ */
+function generateTextFieldSchema(): Record<string, any> {
+  const schema: Record<string, any> = {};
+  for (const langKey of Object.values(LanguageEnum)) {
+    schema[langKey] = { type: String, required: true };
+  }
+  return schema;
+}
 
 @Schema()
 export class NotificationSession {
@@ -12,17 +25,20 @@ export class NotificationSession {
   @Expose()
   @Prop({
     type: String,
-    default: function generateUUID() {
-      return randomUUID();
-    },
+    default: () => randomUUID(),
   })
   _id: string;
 
-  @ApiProperty()
-  @Prop({ default: null })
-  text: string;
+  @ApiProperty({
+    type: Object,
+  })
+  @Prop({
+    type: generateTextFieldSchema(),
+    required: true,
+  })
+  text: Record<LanguageEnum, string>;
 
-  @Prop({ type: Date, expires: 604800, default: Date.now }) //  7 day live for history
+  @Prop({ type: Date, expires: 604800, default: Date.now }) // 7 дней TTL
   expireAt: Date;
 }
